Fix failure when ELSA's cli.pl returns a warning
When ELSA's /opt/elsa/web/cli.pl returns a warning - it prints to
STDOUT. This throws off the position in the array of th response data is
causing the $sensor variable never to be populated and capme to fail
with the - "Failed to find a matching sid..." error message because
the $sensor variable is empty.

This patch will iterate through the response until we find the header
line for the response - and take the next line as the response_data.